memopcombiner: infer offsets of fixed object slots
when this pass runs, all information about the sizes and locations of fixed
objects (negative frame indexes) will be available, so we can determine which
ones are disjoint and can thus be clustered.

note that stack objects (non-negative frame indexes) do not partake in
clustering and are now considered to always alias other memory ops.